Capture and save configuration folders for both directions when defining landscape.
When CCF, QC, SFEE, CEE properties are edited (these options/dialogs will soon be
combined into a single multi-page editor), automatically update property files in both locations.
In the unusual case that the property values have to be different for each direction,
the files will have to be edited manually (outside the tool).

git-svn-id: https://ctf.open.collab.net/svn/repos/ccf/branches/GUIConsole@2427 6771e1e2-0f46-0410-b5d7-d846ff89f743